[Oracle] Initial ADE support; Breaking changes are summarized as follows
1. The metadata table "OBJECTCLASS" is extended to support for
registering new ADE classes along with their inheritance hierarchies and
table mapping information.

2. A couple of additional metadata tables are added for the registration
of arbitrary CityGML ADE schemas. The associations between the
registered ADE schemas and their classes can also be represented.

3. Each CityGML class's table has been extended by adding an
"OBJECTCLASS_ID" column if the column does not exist. The idea of this
implementation is that we want to reuse the existing CityGML tables to
store the data content of ADE objects. This column can hence be used for
identifying the ADE classes which are the subtypes of the respective
CityGML class.

4.Revised JDeveloper database diagrams: The existing metadata tables
"DATABASE_SRS" and "OBJECTCLASS" together with the newly added ones are
grouped to a new separate database diagram named as "Metadata".